requirements.txt
streamlit>=1.31.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
tenacity>=8.2.0
httpx>=0.25.0
firebase-admin>=6.5.0
